+++ savant/src/primitives/point.rs
@@ -8,9 +8,9 @@ use serde_json::Value;
 #[archive(check_bytes)]
 pub struct Point {
     #[pyo3(get, set)]
-    pub x: f64,
+    pub x: f32,
     #[pyo3(get, set)]
-    pub y: f64,
+    pub y: f32,
 }
 
 impl ToSerdeJsonValue for Point {
@@ -36,7 +36,7 @@ impl Point {
     }
 
     #[new]
-    pub fn new(x: f64, y: f64) -> Self {
+    pub fn new(x: f32, y: f32) -> Self {
         Self { x, y }
     }
 }
